Update project, but it does not compile

+++ src/main/java/ru/skypro/homework/controller/AnnounceController.java
@@ -79,7 +79,7 @@ public class AnnounceController {
                     })
     })
     @GetMapping("{id]")
-    public ResponseEntity<?> get(@PathVariable Long id) {
+    public ResponseEntity<?> get(@PathVariable Integer id) {
         return ResponseEntity.status(HttpStatus.OK).body(announceService.get(id));
     }
 
@@ -99,7 +99,7 @@ public class AnnounceController {
                     })
     })
     @DeleteMapping("{id]")
-    public ResponseEntity<?> delete(@PathVariable Long id) {
+    public ResponseEntity<?> delete(@PathVariable Integer id) {
         announceService.delete(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
@@ -121,7 +121,7 @@ public class AnnounceController {
                     })
     })
     @PatchMapping("{id]")
-    public ResponseEntity<?> updateInfo(@PathVariable Long id,
+    public ResponseEntity<?> updateInfo(@PathVariable Integer id,
                                         @RequestParam CreateOrUpdateAd property) {
         return ResponseEntity.status(HttpStatus.OK).body(announceService.updateInfo(id, property));
     }
@@ -160,7 +160,7 @@ public class AnnounceController {
                     })
     })
     @PatchMapping(value = "{id]/image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<?> updateImage(@PathVariable Long id,
+    public ResponseEntity<?> updateImage(@PathVariable Integer id,
                                          @RequestPart MultipartFile image) {
         try {
             announceService.updateImage(id, image);
